Updated CONTRIBUTING.MD
Added a testing section to the CONTRIBUTING.MD file

Added more testing

Added more testing functions to check the core of the program when converting text or markdown files to html.

Added NUnit for testing

Added the NUnit framework for testing purposes. Created a new project for testing. Added test methods to check the behavior of OutputDirectoryHandler with different scenarios.